+++ src/list/slice.test.ts
@@ -72,3 +72,15 @@ type Slice_Spec = [
   // @ts-expect-error
   $<$<$<$<List.Slice, 1>, 2>, 3>, [1, 2, 3]>
 ]
+
+it('should return a slice of a list', () => {
+  expect(List.slice(1)(3)([1, 2, 3, 4, 5])).toEqual([2, 3])
+})
+
+it('should return a slice of a list with negative indices', () => {
+  expect(List.slice(-2)(-1)([1, 2, 3, 4, 5])).toEqual([4])
+})
+
+it('can handle empty lists', () => {
+  expect(List.slice(0)(0)([])).toEqual([])
+})